FLPathSegment addClippingsWithIntersections:info:isFirst: check
When generating intersections for lines, start and end points should
be disregarded, because they don't intersect in fact.

Later the intersection function for line segments has to keep the
t (time) itself, then this is not so much out of place any longer.

This problem came up with overlapping squares on the same
height. There are still other problems, namely with line segments that
are on top of each other. One segment is doubled, the other is
missing.